feat(kustomize): Add steps to deploy Spinnaker with Spinnaker
We need to add `app.kubernetes.io/managed-by: Spinnaker` label to reduce
redeployment of all pods when Spinnaker pipeline executes.
The redeployment disrupts MariaDB and prevents pipeline completion.
A second pipeline execution succeeds as a no-op.

The first deployment also adds annotations to Kubernetes resources so it
is possible to verify it has run.